% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20016R.R
\name{lap_fun}
\alias{lap_fun}
\title{Laplace funcrion}
\usage{
lap_fun(x)
}
\arguments{
\item{x}{the independent variable of laplace function}
}
\value{
the value of laplace funcrion
}
\description{
compute the value of laplace function
}
