% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20016R.R
\name{max_extremeout}
\alias{max_extremeout}
\title{maximum number of extreme points}
\usage{
max_extremeout(x, y)
}
\arguments{
\item{x}{the first sample}

\item{y}{the second sample}
}
\value{
the maximum number of extreme points
}
\description{
maximum number of extreme points of each sample with respect to the other sample
}
