% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIrayleigh.R
\name{MCIrayleigh}
\alias{MCIrayleigh}
\title{Monte carlo integration of the probability density function of rayleigh distribution}
\usage{
MCIrayleigh(a, b, sig, n)
}
\arguments{
\item{a}{the lower bound of integration}

\item{b}{the upper bound of integration}

\item{sig}{the parameter of rayleigh distribution}

\item{n}{the number of random uniform distributed samples}
}
\value{
the value of integration
}
\description{
Using a simple algorithm to compute integration of the PDF of rayleigh distribution
}
