% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrayleigh.R
\name{rrayleigh}
\alias{rrayleigh}
\title{Generate random numbers from a Rayleigh distribution}
\usage{
rrayleigh(n, sig)
}
\arguments{
\item{n}{the number of samples}

\item{sig}{the parameter of rayleigh distribution}
}
\value{
a random sample from of rayleigh distribution size \code{n}
}
\description{
Using inverse transformation method to generate rayleigh distributed samples
}
\examples{
\dontrun{
x <- rrayleigh(10,2)
}
}
