% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20016R.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{Random walk sampling}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the parameter of varience}

\item{x0}{the initial value}

\item{N}{the number of simulation}
}
\value{
the random samples from laplace distribution
}
\description{
Use random walk sampling to generate samples from laplace distribution
}
