\name{StatComp20016-package}
\alias{StatComp20016-package}
\docType{package}
\title{
  R and Rcpp functions of homework for 'Statistical Computing' course
}
\description{
  R functions to generate random numbers from rayleigh distribution and to compute monte carlo integration are provided in this package.
}
\details{
  This section provide a more detailed overview of how to use the package, including the most important functions.
}
\author{
Hongyi Liao

Maintainer: Hongyi Liao <1013475972@qq.com>
}
\keyword{ package }
